Feat: Adiciona Mapper e DTO, ajusta Service e Controller para a Entidade Proposicao

+++ src/main/java/com/example/Deputados/Proposicao/ProposicaoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
-public interface ProposicaoRepository extends JpaRepository<ProposicaoModel, Integer>{
+public interface ProposicaoRepository extends JpaRepository<ProposicaoModel, Long>{
     
 }
